% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mplot_for_multiple_QTL_GWAS_intersect.R
\name{Mplot_for_multiple_QTL_GWAS_intersect}
\alias{Mplot_for_multiple_QTL_GWAS_intersect}
\title{Mplot for three type of QTLs with GWAS association
MarkerName : eg. 1:10086:A/C or 1:10086:A:C
Usage eg. Mplot_for_multiple_QTL_GWAS_intersect('/A/B/C','a.TBL','QTL1.gz','QTL2.gz,'QTL3.gz,'/a/b/c','plotA',5e-04,10,1,5.5,20)}
\usage{
Mplot_for_multiple_QTL_GWAS_intersect(
  file_path,
  file_name,
  qtl_file1,
  qtl_file2,
  qtl_file3,
  figure_pathy,
  figure_name,
  p_col,
  snp_col,
  p_threshold,
  height,
  width
)
}
\arguments{
\item{file_path}{the pathway of meta file}

\item{file_name}{the name of of meta file with suffix}

\item{qtl_file1}{the file of type 1 QTL}

\item{qtl_file2}{the file of type 2 QTL}

\item{qtl_file3}{the file of type 3 QTL}

\item{figure_pathy}{the pathway to save Mplot}

\item{figure_name}{the name of of Mplot}

\item{p_col}{column number of p-value}

\item{snp_col}{colnum number of MarkerName}

\item{p_threshold}{P-value threshold of GWAS association}

\item{height}{height of Mplot}

\item{width}{width of Mplot}
}
\description{
Mplot for three type of QTLs with GWAS association
MarkerName : eg. 1:10086:A/C or 1:10086:A:C
Usage eg. Mplot_for_multiple_QTL_GWAS_intersect('/A/B/C','a.TBL','QTL1.gz','QTL2.gz,'QTL3.gz,'/a/b/c','plotA',5e-04,10,1,5.5,20)
}
